+++ plugins/toolbar/plugin.js
@@ -415,9 +415,16 @@
 	} );
 
 	function getToolbarConfig( editor ) {
-		var removeButtons = editor.config.removeButtons;
+		var removeButtons = getRemoveButtons( editor.config.removeButtons );
 
-		removeButtons = removeButtons && removeButtons.split( ',' );
+		// (#5122)
+		function getRemoveButtons( config ) {
+			if ( config && typeof config === 'string' ) {
+				return config.split( ',' );
+			}
+
+			return config;
+		}
 
 		function buildToolbarConfig() {
 
@@ -771,7 +778,15 @@ CKEDITOR.config.toolbarLocation = 'top';
  * List of toolbar button names that must not be rendered. This will also work
  * for non-button toolbar items, like the Font drop-down list.
  *
- *		config.removeButtons = 'Underline,JustifyCenter';
+ * ```javascript
+ * config.removeButtons = 'Underline,JustifyCenter';
+ * ```
+ *
+ * Since version 4.20.0 you can also pass an array of button names:
+ *
+ * ```javascript
+ * config.removeButtons = [ 'Underline', 'JustifyCenter' ];
+ * ```
  *
  * This configuration option should not be overused. The recommended way is to use the
  * {@link CKEDITOR.config#removePlugins} setting to remove features from the editor
@@ -780,7 +795,7 @@ CKEDITOR.config.toolbarLocation = 'top';
  * In some cases though, a single plugin may define a set of toolbar buttons and
  * `removeButtons` may be useful when just a few of them are to be removed.
  *
- * @cfg {String} [removeButtons]
+ * @cfg {String/String[]} [removeButtons]
  * @member CKEDITOR.config
  */